add abstract class for client_repository.py
